fix demo pages
